Fixed a small memory leak.
git-svn-id: http://svn.osgeo.org/geos/trunk@211 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ source/headers/indexQuadtree.h
@@ -50,7 +50,8 @@ public:
 	string toString();
 private:
 	double x;
-	long long xBits;
+//	long long xBits;
+	long xBits;
 };
 
 /**
