+++ source/index/quadtree/DoubleBits.cpp
@@ -6,8 +6,10 @@ namespace geos {
 double DoubleBits::powerOf2(int exp){
 	if (exp>1023 || exp<-1022)
 		throw new IllegalArgumentException("Exponent out of bounds");
-	long long expBias=exp+EXPONENT_BIAS;
-	long long bits=(long long)expBias << 52;
+//	long long expBias=exp+EXPONENT_BIAS;
+//	long long bits=(long long)expBias << 52;
+	long expBias=exp+EXPONENT_BIAS;
+	long bits=(long)expBias << 52;
 	return (double) bits;
 }
 
@@ -40,7 +42,8 @@ double DoubleBits::maximumCommonMantissa(double d1, double d2) {
 
 DoubleBits::DoubleBits(double nx) {
 	x=nx;
-	xBits=(long long)x;
+//	xBits=(long long)x;
+	xBits=(long)x;
 }
 
 double DoubleBits::getDouble() {
@@ -53,7 +56,8 @@ double DoubleBits::getDouble() {
 * @return
 */
 int DoubleBits::biasedExponent(){
-	long long signExp=(long long)(xBits>>52);
+//	long long signExp=(long long)(xBits>>52);
+	long signExp=(long)(xBits>>52);
 	int exp=signExp&0x07ff;
 	return exp;
 }
